fix: move MachineTranslationTest to common packages to use it for slack
